Set up localization
...with initial support for English and Spanish. We use Python's builtin
`gettext()` facilities to localize messages in templates, and the new
`localize` script helps automate the process of extracting messages to
be translated and compiling them. Translations are placed under the
`locale` directory at the root of the application, following the usual
GNU gettext structure.

Presently, we expect all translatable messages to occur in `views/*.tpl`
files, and `xgettext`, the tool that `localize` uses to extract
messages, will only reliably extract messages on pure-python lines (ones
that begin with `%` and contain only Python code).

`localize` expects you to have the `gettext` utilities available in your
PATH. In particular, it currently uses the `xgettext`, `msgmerge`, and
`msgfmt` utilities.