Multiple Choice: Update app for validation
This is basically a new app because the user interface post-login has
changed completely to support validating transcriptions. Instead of
adding a single annotation per recording, there are now a number of
select boxes associated with each recording representing different
aspects of the quality of the recording and the transcription. There is
also a final "Checked" select that, when set to 1, treats the file as
"checked" and sets a "checked_at" time stamp. If "Checked" is set to
0 then any existing "checked_at" time stamp is cleared.

The navigation has also changed significantly. On each recording, there
are links to move to the directly-preceding recording, the
directly-following recording, the first unchecked recording before the
current recording ("Skip Prev") and the first unchecked recording after
the current recording ("Skip Next"). The "Skip" links will wrap around
while the direct links will not.

What was previously the "Next" button has become the "Save" button, and
it's associated with a "mode", one of  "skip prev", "prev", "next", or
"skip next". Clicking the "Save" button submits any changes for the
recording and, after saving them to the DB, uses the mode to determine
which file to go to next according to the same logic as the navigation
links, passing a "mode" query parameter in order to preserve the
current mode.

There are also a number of HTML/CSS changes to (hopefully) make the app
a little prettier and better suited to the new workflow.