Multiple Choice: add current mode to nav links
With this change, clicking navigation links will act as a "one time"
operation, otherwise maintaining the current mode. So, for example, if
the current mode is "skip next", clicking the "skip prev" navigation
link will move to the previous unchecked file, but the mode will still
be "skip next", so saving without changing the mode will skip forward to
the next unchecked file.

The only way to change the mode via the UI is to change the dropdown and
save the current file. Since the mode is pulled from a query param
however, you _can_ change the mode by manually appending, e.g.,
`?mode=skip_next` to a `/annotate/{id}` URL.